backend/pyproject.toml
[tool.poetry]
name = "academie prologin"
version = "0.1.0"
description = ""
authors = ["Marc 'risson' Schmitt <marc.schmitt@prologin.org>", "Association Prologin <info@prologin.org>"]
license = "AGPL-3.0"

[[tool.poetry.source]]
name = "gitlab-prologin"
url = "https://gitlab.com/api/v4/groups/1331525/-/packages/pypi/simple"
secondary = true

[tool.black]
line-length = 79

[tool.poetry.dependencies]
python = "^3.10"
boto3 = "^1.24"
collectfast = "*"
coverage = { version = "*", extras = [ "toml" ] }
Django = "^4.0.2"
celery = "^5.2.3"
redis = "^4.1.4"
Jinja2 = "^3.0.3"
djangorestframework = "^3.13.1"
django-filter = "^21.1"
django-debug-toolbar = "*"
django-prometheus = "*"
django-storages = "^1.12"
gunicorn = "^20.1"
ipython = "*"
prologin-django-utils = "0.0.1"
prospector = {extras = ["with_everything"], version = "*"}
psycopg2 = "^2.9"
requests = "^2.27.1"
python-dateutil = "^2.8.2"
drf-yasg = "^1.20.0"
djangorestframework-simplejwt = "^5.1.0"
django-rest-passwordreset = "^1.2.1"
django-resized = "^1.0.0"
Pillow = "^9.1.1"

[tool.poetry.dev-dependencies]
pre-commit = "^2.17.0"
black = "^22.1.0"
coverage = "^6.3.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma=true


[tool.coverage.run]
branch = true
omit = [
    "../*migrations*",
]

[tool.coverage.report]
show_missing = true
exclude_lines = [
    # Re-enable the standard pragma
    "pragma: nocover",

    # Don't complain if tests don't hit defensive assertion code:
    "raise NotImplementedError",

    # Don't complain about missing debug-only code:
    "def __repr__",

    # Don't complain if non-runnable code isn't run:
    "if 0:",
    "if __name__ == .__main__.:",
]
